fix issues found by lint

+++ os/osutil/copyandmovefile_linux.go
@@ -1,4 +1,4 @@
-// +build linux
+//go:build linux
 
 package osutil
 
@@ -24,5 +24,5 @@ func getExtendedFileInfo(fi fs.FileInfo) *extendedFileInfo {
 }
 
 func timespecToTime(ts syscall.Timespec) time.Time {
-	return time.Unix(int64(ts.Sec), int64(ts.Nsec))
+	return time.Unix(ts.Sec, ts.Nsec)
 }
